services/tldr/requirements.txt
jsonschema>=4.23.0 # for validation
pika>=1.3.2 # AMQP client
logging>=0.4.9.6 # logger
python-dotenv>=1.0.1 # environemnt variables

# Model Dependencies
huggingface_hub[cli]==0.27.1
llama_cpp_agent==0.2.35